feat: customisation updates
