+++ apps/agent/entrypoints/options/customization/CustomizationPage.tsx
@@ -1,10 +1,12 @@
 import type { FC } from 'react'
-import { CustomizationCard } from './CustomizationCard'
+import { CustomizationHeader } from './CustomizationHeader'
+import { ToolbarSettingsCard } from './ToolbarSettingsCard'
 
 export const CustomizationPage: FC = () => {
   return (
     <div className="fade-in slide-in-from-bottom-5 animate-in space-y-6 duration-500">
-      <CustomizationCard />
+      <CustomizationHeader />
+      <ToolbarSettingsCard />
     </div>
   )
 }
